// src/interfaces/MediaItemDetails.ts
export interface IMediaItemDetail {
    adult:                 boolean;
    backdrop_path:         string | null;
    created_by?:           CreatedBy[];
    episode_run_time?:     number[];
    first_air_date?:       string;
    genres:                Genre[];
    homepage:              string;
    id:                    number;
    in_production?:        boolean;
    languages?:            string[];
    last_air_date?:        string;
    last_episode_to_air?:  LastEpisodeToAir;
    name?:                 string;
    next_episode_to_air?:  null;
    networks?:             Network[];
    number_of_episodes?:   number;
    number_of_seasons?:    number;
    origin_country?:       string[];
    original_language:     string;
    original_name?:        string;
    overview:              string;
    popularity:            number;
    poster_path:           string;
    production_companies?: any[];
    production_countries?: any[];
    seasons?:              Season[];
    spoken_languages:      SpokenLanguage[];
    status?:               string;
    tagline?:              string;
    type?:                 string;
    vote_average:          number;
    vote_count:            number;
    belongs_to_collection?: BelongsToCollection;
    budget?:               number;
    imdb_id?:              string;
    original_title?:       string;
    release_date?:         string;
    revenue?:              number;
    runtime?:              number;
    video?:                boolean;
    title?:                string;
}

export interface BelongsToCollection {
    id:            number;
    name:          string;
    poster_path:   string;
    backdrop_path: string;
}

export interface Genre {
    id:   number;
    name: string;
}

export interface ProductionCompany {
    id:             number;
    logo_path:      null | string;
    name:           string;
    origin_country: string;
}

export interface ProductionCountry {
    iso_3166_1: string;
    name:       string;
}

export interface SpokenLanguage {
    english_name: string;
    iso_639_1:    string;
    name:         string;
}

export interface CreatedBy {
    id:           number;
    credit_id:    string;
    name:         string;
    gender:       number;
    profile_path: string;
}

export interface Genre {
    id:   number;
    name: string;
}

export interface LastEpisodeToAir {
    id:              number;
    name:            string;
    overview:        string;
    vote_average:    number;
    vote_count:      number;
    air_date:        string;
    episode_number:  number;
    production_code: string;
    runtime:         number;
    season_number:   number;
    show_id:         number;
    still_path:      null;
}

export interface Network {
    id:             number;
    logo_path:      string;
    name:           string;
    origin_country: string;
}

export interface Season {
    air_date:      string;
    episode_count: number;
    id:            number;
    name:          string;
    overview:      string;
    poster_path:   string;
    season_number: number;
}

export interface SpokenLanguage {
    english_name: string;
    iso_639_1:    string;
    name:         string;
}